`tests/array_api/` docstrings and re-org
