+++ hypothesis-python/tests/array_api/test_argument_validation.py
@@ -229,6 +229,7 @@ def test_raise_invalid_argument(xp, xps, strat_name, kwargs):
 
 
 @pytest.mark.parametrize("api_version", [..., "latest", "1970.01", 42])
-def test_make_namespace_raise_invalid_argument(xp, api_version):
+def test_make_strategies_namespace_raise_invalid_argument(xp, api_version):
+    """Function raises helpful error with invalid arguments."""
     with pytest.raises(InvalidArgument):
         make_strategies_namespace(xp, api_version=api_version)
